Update save file img when user upload
