+++ IServices/IFeedbackService.cs
@@ -1,0 +1,7 @@
+﻿namespace booking.IServices
+{
+    public interface IFeedbackService
+    {
+        bool isTrueIMG(IFormFile img);
+    }
+}
