+++ Views/Home/Index.cshtml
@@ -474,45 +474,45 @@
 
           <div class="col-lg-8 mt-5 mt-lg-0">
 
-            <form asp-controller="Home" asp-action="createFeedback" method="post" role="form" class="php-email-form" id="feedbackForm">
+            <form asp-controller="Home" asp-action="createFeedback" method="post" role="form" class="php-email-form" id="feedbackForm" enctype="multipart/form-data">
               <div class="row">
                 <div class="col-md-6 form-group">
-                  <input type="text" name="name" class="form-control" id="name" placeholder="Your Name" required>
+                    <input type="text" name="name" class="form-control" id="name" placeholder="Your Name" required>
                 </div>
                 <div class="col-md-6 form-group mt-3 mt-md-0">
-                  <input type="text" class="form-control" name="jobs" id="jobs" placeholder="Your Jobs" required>
+                    <input type="text" class="form-control" name="jobs" id="jobs" placeholder="Your Jobs" required>
                 </div>
               </div>
-              <div class="row">
-                <div class="col-md-6 form-group">
-                  <input type="email" name="email" class="form-control" id="email" placeholder="Your Email" required>
+               <div class="row">
+                 <div class="col-md-6 form-group">
+                   <input type="email" name="email" class="form-control" id="email" placeholder="Your Email" required>
+                 </div>
+                 <div class="col-md-6 form-group mt-3 mt-md-0">
+                   <input type="number" class="form-control" name="phone" id="phone" placeholder="Your Phone" required>
+                 </div>
+               </div>
+               <div class="form-group mt-3 file-input-wrapper" style="position: relative;">
+                  <input type="file" class="form-control" name="img" id="img">
+                  <label for="img" class="custom-file-label" style="position: absolute;
+                    top: 0;
+                    left: 0;
+                    width: 100%;
+                    height: 100%;
+                    pointer-events: none;
+                    padding: 0.375rem 0.75rem;
+                    background-color:black;
+                    border: 1px solid #ced4da;
+                    border-radius: 0.25rem;">May you upload your photo? please click here to upload. Thanks</label>
+               </div>
+               <div class="form-group mt-3">
+                  <textarea class="form-control" id="feedback" name="feedback" rows="8" placeholder="Your feedback" required></textarea>
+               </div>
+               <div class="my-3">
+                  <div class="loading">Loading</div>
+                  <div class="error-message"></div>
+                  <div class="sent-message">Your message has been sent. Thank you!</div>
                 </div>
-                <div class="col-md-6 form-group mt-3 mt-md-0">
-                  <input type="number" class="form-control" name="phone" id="phone" placeholder="Your Phone" required>
-                </div>
-              </div>
-              <div class="form-group mt-3 file-input-wrapper" style="position: relative;">
-                <input type="file" class="form-control" name="img" id="img">
-                <label for="img" class="custom-file-label" style="position: absolute;
-                  top: 0;
-                  left: 0;
-                  width: 100%;
-                  height: 100%;
-                  pointer-events: none;
-                  padding: 0.375rem 0.75rem;
-                  background-color:black;
-                  border: 1px solid #ced4da;
-                  border-radius: 0.25rem;">May you upload your photo? please click here to upload. Thanks</label>
-              </div>
-              <div class="form-group mt-3">
-                <textarea class="form-control" id="feedback" name="feedback" rows="8" placeholder="Your feedback" required></textarea>
-              </div>
-              <div class="my-3">
-                <div class="loading">Loading</div>
-                <div class="error-message"></div>
-                <div class="sent-message">Your message has been sent. Thank you!</div>
-              </div>
-              <div class="text-center"><button id="button-feedback" type="submit">Give Feedback</button></div>
+                <div class="text-center"><button id="button-feedback" type="submit">Give Feedback</button></div>
             </form>
             <div id="thankYouMessage" style="display: none; position: fixed; top: 0; left: 0; width: 100%; height: 100%; background: rgba(0, 0, 0, 0.8); color: white; text-align: center; padding-top: 20%; font-size: 24px;">
                 Thank you for your feedback!
